Move remaining hash utils to legacy
