+++ cpp/src/hash/legacy/hash_functions.cuh
