+++ cpp/src/hash/legacy/hashing.cu
@@ -21,7 +21,7 @@
 #include <cudf/utilities/error.hpp>
 #include <join/legacy/joining.h>
 #include <table/legacy/device_table.cuh>
-#include "hash/hash_functions.cuh"
+#include "hash/legacy/hash_functions.cuh"
 #include <utilities/int_fastdiv.h>
 #include <cudf/utilities/nvtx_utils.hpp>
 #include <copying/legacy/scatter.hpp>
