+++ cpp/src/hash/legacy/helper_functions.cuh
