+++ cpp/src/hash/legacy/managed.cuh
