+++ cpp/src/hash/legacy/managed_allocator.cuh
