+++ cpp/src/table/legacy/device_table.cuh
@@ -22,8 +22,8 @@
 #include <utilities/cudf_utils.h>
 #include <bitmask/legacy/bit_mask.cuh>
 #include <bitmask/legacy/legacy_bitmask.hpp>
-#include <hash/hash_functions.cuh>
-#include <hash/managed.cuh>
+#include <hash/legacy/hash_functions.cuh>
+#include <hash/legacy/managed.cuh>
 #include <cudf/legacy/table.hpp>
 #include <cudf/utilities/error.hpp>
 #include <cudf/utilities/legacy/type_dispatcher.hpp>
