+++ cpp/tests/hash_map/legacy/multimap_test.cu
@@ -19,7 +19,7 @@
 #include <thrust/device_vector.h>
 
 #include <cudf/cudf.h>
-#include <hash/concurrent_unordered_multimap.cuh>
+#include <hash/legacy/concurrent_unordered_multimap.cuh>
 
 #include <gtest/gtest.h>
